chore: update dependencies in go.mod and go.sum
The versions of some dependencies of the project have been updated. Specifically, the versions of "github.com/gowool/extends-template" and "github.com/roadrunner-server/endure/v2" were updated to their newer versions. This was necessary to incorporate new features and bug fixes provided in these libraries. In addition, a new version of "golang.org/x/exp" is used now for the same reasons.